ASL/requirements.txt
# Strands Agents Framework
strands-agents>=1.0.0

# AWS Bedrock AgentCore
bedrock-agentcore>=1.0.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Additional dependencies
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0
